requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
huggingface_hub
